app/requirements.txt
# Backend
requests
django
WeasyPrint # Generate PDFs for email

# Infrastructure
psycopg2-binary
gunicorn
sentry-sdk
django-q

# API
djangorestframework
django-cors-headers

# Static files
whitenoise

# Admin
django-extensions
ipython

# Phone calls
twilio

# Development
watchdog[watchmedo]

# File storage
boto3
django-storages
Pillow

# Linting
isort
black
flake8

# Testing
faker
pytest-django
pytest-factoryboy
# https://github.com/Teemu/pytest-sugar/issues/187
# pytest-sugar
responses
freezegun

# Admin
pygments  # Pretty print JSON in admin